취합 완료

+++ src/main/java/com/green/backend/application/dto/ApplicationDTO.java
@@ -15,7 +15,7 @@ import java.time.LocalDateTime;
 public class ApplicationDTO {
     private Long detailId;              // 답사번호
     private Long memberId;           // 기업번호
-    private Integer expertId;           // 전문가번호
+    private Long expertId;           // 전문가번호
 
     private Integer times;              // 정기차수
     private String surveyStatus;        // 상태 (신청/진행중/완료)
